+++ src/code/_.rs
@@ -28,6 +28,7 @@ pub mod util; // Cross-cutting code and macro utilities
 
 util::mods_in! {
     pub mod_ any; // Dynamic typing, type identity, and type inspection
+        mod_ build; // Build-process support and integration
     pub mod_ convert; // Type conversion and adaptation
     pub mod_ hint; // Compiler and execution hints
     pub mod_ init; // Default and const-friendly initialization
@@ -37,7 +38,12 @@ util::mods_in! {
     pub mod_ result; // Generic outcome and resolution types
     pub mod_ source; // Source-code location, provenance, and inclusion
 }
-util::mods_out! { // _pub_mods, _reexports, _crate_internals
+util::mods_out! { // _mods, _pub_mods, _reexports, _crate_internals
+    _mods {
+        pub use super::{
+            build::_all::*,
+        };
+    }
     _pub_mods {
         pub use super::{
             any::_all::*,