+++ docs/changelog.md
@@ -55,6 +55,7 @@
 
 ## code
 - new modules: `any`, `convert`, `hint`, `init`, `source`.
+- move the root `build` module here.
 
 ### code::marker
 - fix `IndexRepr` impl for `usize`.